+++ lib/l10n/arb/app_en.arb
@@ -2,6 +2,13 @@
     "@@locale": "en",
     "counterAppBarTitle": "Counter",
     "@counterAppBarTitle": {
-        "description": "カウンターページのAppBarに表示されるテキスト"
-    }
+        "description": "Texto mostrado en la AppBar de la página del contador"
+    },
+    "travel" :"Travel",
+    "food": "Ẩm thực",
+    "seeAll" : "See All",
+    "interestingSpaces" : " Interesting places",
+    "exploration":"Exploration",
+    "whereDoYouWantToGo" : "Where Do You Want To Go",
+    "seeMore" : "See More"
 }
