+++ lib/l10n/arb/app_vi.arb
@@ -1,0 +1,14 @@
+{
+    "@@locale": "vi",
+    "counterAppBarTitle": "Counter",
+    "@counterAppBarTitle": {
+        "description": "Texto mostrado en la AppBar de la página del contador"
+    },
+    "travel" :"'Travel'",
+    "food": "Ẩm thực",
+    "seeAll" : "Xem tất cả",
+    "interestingSpaces" : "Các điểm lý thú",
+    "explorationTourism":"Khám Phá",
+    "whereDoYouWantToGo" : "Bạn muốn đi đâu",
+    "seeMore" : "Xem Thêm"
+}
