--- lib/presentation/feature/home/home_page_tab.dart
@@ -1,4 +1,0 @@
-enum HomePageTab {
-  todos,
-  profile,
-}
